requirements.txt
fastapi
uvicorn
mysql-connector-python
gspread
google-auth-oauthlib